// src/entities/Article/model/consts/articleConsts.ts
export enum ArticleBlockType {
  CODE = 'CODE',
  IMAGE = 'IMAGE',
  TEXT = 'TEXT',
}

export enum ArticleSortField {
  CREATED = 'createdAt',
  TITLE = 'title',
  VIEWS = 'views',
}

export enum ArticleType {
  ALL = 'ALL',
  ECONOMICS = 'ECONOMICS',
  IT = 'IT',
  SCIENCE = 'SCIENCE',
}

export enum ArticleView {
  LIST = 'list', // BIG
  PLATE = 'plate', // SMALL
}